# docs/swagger.yaml
basePath: /api
definitions:
  controller.KlingCameraConfig:
    properties:
      horizontal:
        example: 2.5
        type: number
      pan:
        example: 0
        type: number
      roll:
        example: 0
        type: number
      tilt:
        example: 0
        type: number
      vertical:
        example: 0
        type: number
      zoom:
        example: 0
        type: number
    type: object
  controller.KlingCameraControl:
    properties:
      config:
        $ref: '#/definitions/controller.KlingCameraConfig'
      type:
        example: simple
        type: string
    type: object
  controller.KlingImage2VideoRequest:
    properties:
      aspect_ratio:
        example: "16:9"
        type: string
      callback_url:
        example: https://your.domain/callback
        type: string
      camera_control:
        $ref: '#/definitions/controller.KlingCameraControl'
      cfg_scale:
        example: 0.7
        type: number
      duration:
        example: "5"
        type: string
      external_task_id:
        example: custom-task-002
        type: string
      image:
        example: https://h2.inkwai.com/bs2/upload-ylab-stunt/se/ai_portal_queue_mmu_image_upscale_aiweb/3214b798-e1b4-4b00-b7af-72b5b0417420_raw_image_0.jpg
        type: string
      mode:
        example: std
        type: string
      model_name:
        example: kling-v2-master
        type: string
      negative_prompt:
        example: blurry, low quality
        type: string
      prompt:
        example: A cat playing piano in the garden
        type: string
    required:
    - image
    type: object
  controller.KlingText2VideoRequest:
    properties:
      aspect_ratio:
        example: "16:9"
        type: string
      callback_url:
        example: https://your.domain/callback
        type: string
      camera_control:
        $ref: '#/definitions/controller.KlingCameraControl'
      cfg_scale:
        example: 0.7
        type: number
      duration:
        example: "5"
        type: string
      external_task_id:
        example: custom-task-001
        type: string
      mode:
        example: std
        type: string
      model_name:
        example: kling-v1
        type: string
      negative_prompt:
        example: blurry, low quality
        type: string
      prompt:
        example: A cat playing piano in the garden
        type: string
    required:
    - prompt
    type: object
  controller.SupplierApplicationReviewRequest:
    properties:
      reason:
        type: string
      status:
        type: integer
    type: object
  controller.SupplierApplicationSubmitRequest:
    properties:
      business_license_url:
        type: string
      company_name:
        type: string
      company_size:
        type: string
      contact_mobile:
        type: string
      contact_name:
        type: string
      contact_wechat:
        type: string
      credit_code:
        type: string
      legal_representative:
        type: string
    type: object
  dto.VideoRequest:
    properties:
      duration:
        description: Video duration (seconds)
        example: 5
        type: number
      fps:
        description: Video frame rate
        example: 30
        type: integer
      height:
        description: Video height
        example: 512
        type: integer
      image:
        description: Image input (URL/Base64)
        example: https://h2.inkwai.com/bs2/upload-ylab-stunt/se/ai_portal_queue_mmu_image_upscale_aiweb/3214b798-e1b4-4b00-b7af-72b5b0417420_raw_image_0.jpg
        type: string
      metadata:
        additionalProperties: {}
        description: Vendor-specific/custom params (e.g. negative_prompt, style, quality_level,
          etc.)
        type: object
      model:
        description: Model/style ID
        example: kling-v1
        type: string
      "n":
        description: Number of videos to generate
        example: 1
        type: integer
      prompt:
        description: Text prompt
        example: 宇航员站起身走了
        type: string
      response_format:
        description: Response format
        example: url
        type: string
      seed:
        description: Random seed
        example: 20231234
        type: integer
      user:
        description: User identifier
        example: user-1234
        type: string
      width:
        description: Video width
        example: 512
        type: integer
    type: object
  dto.VideoTaskError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  dto.VideoTaskMetadata:
    properties:
      duration:
        description: 实际生成的视频时长
        example: 5
        type: number
      fps:
        description: 实际帧率
        example: 30
        type: integer
      height:
        description: 实际高度
        example: 512
        type: integer
      seed:
        description: 使用的随机种子
        example: 20231234
        type: integer
      width:
        description: 实际宽度
        example: 512
        type: integer
    type: object
  dto.VideoTaskResponse:
    properties:
      error:
        allOf:
        - $ref: '#/definitions/dto.VideoTaskError'
        description: 错误信息（失败时）
      format:
        description: 视频格式
        example: mp4
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/dto.VideoTaskMetadata'
        description: 结果元数据
      status:
        description: 任务状态
        example: succeeded
        type: string
      task_id:
        description: 任务ID
        example: abcd1234efgh
        type: string
      url:
        description: 视频资源URL（成功时）
        type: string
    type: object
info:
  contact: {}
  description: TokenFactory backend API documentation powered by swaggo.
  title: TokenFactory API
  version: "1.0"
paths:
  /kling/v1/videos/image2video:
    post:
      consumes:
      - application/json
      description: 调用可灵AI图生视频接口，生成视频内容
      parameters:
      - description: '用户认证令牌 (Aeess-Token: sk-xxxx)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: 图生视频请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.KlingImage2VideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.VideoTaskResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 可灵官方-图生视频
      tags:
      - Video
  /kling/v1/videos/image2video/{task_id}:
    get:
      consumes:
      - application/json
      description: Query the status and result of a Kling video generation task by
        task ID
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 可灵任务查询--图生视频
      tags:
      - Origin
  /kling/v1/videos/text2video:
    post:
      consumes:
      - application/json
      description: 调用可灵AI文生视频接口，生成视频内容
      parameters:
      - description: '用户认证令牌 (Aeess-Token: sk-xxxx)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: 视频生成请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.KlingText2VideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.VideoTaskResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 可灵文生视频
      tags:
      - Video
  /kling/v1/videos/text2video/{task_id}:
    get:
      consumes:
      - application/json
      description: Query the status and result of a Kling text-to-video generation
        task by task ID
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 可灵任务查询--文生视频
      tags:
      - Origin
  /user/messages/{id}/read:
    post:
      parameters:
      - description: 消息ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success + data{updated}
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 标记当前用户消息为已读
      tags:
      - Message
  /user/messages/self:
    get:
      parameters:
      - description: 页码
        in: query
        name: p
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页结果
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查询当前用户站内消息
      tags:
      - Message
  /user/messages/unread_count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: success + data{unread_count}
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户未读站内消息数量
      tags:
      - Message
  /user/supplier/application:
    get:
      parameters:
      - description: 页码
        in: query
        name: p
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 状态：0待审核 1审核通过 2审核驳回
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页结果
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 管理员分页查询供应商申请
      tags:
      - SupplierAdmin
    post:
      consumes:
      - application/json
      description: 普通用户提交供应商申请，提交后生成管理员待审核站内消息
      parameters:
      - description: 申请信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.SupplierApplicationSubmitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success + data{id,status}
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 提交供应商入驻申请
      tags:
      - Supplier
  /user/supplier/application/{id}/review:
    post:
      consumes:
      - application/json
      description: 任一管理员可审核一次，仅待审核状态允许处理
      parameters:
      - description: 申请ID
        in: path
        name: id
        required: true
        type: integer
      - description: 审核信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.SupplierApplicationReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success + data{id,status}
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 管理员审核供应商申请
      tags:
      - SupplierAdmin
  /user/supplier/application/self:
    get:
      parameters:
      - description: 页码
        in: query
        name: p
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 状态：0待审核 1审核通过 2审核驳回
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页结果
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查询当前用户提交的供应商申请
      tags:
      - Supplier
  /v1/video/generations:
    post:
      consumes:
      - application/json
      description: |-
        调用视频生成接口生成视频
        支持多种视频生成服务：
        - 可灵AI (Kling): https://app.klingai.com/cn/dev/document-api/apiReference/commonInfo
        - 即梦 (Jimeng): https://www.volcengine.com/docs/85621/1538636
      parameters:
      - description: '用户认证令牌 (Aeess-Token: sk-xxxx)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: 视频生成请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VideoRequest'
      produces:
      - application/json
      responses:
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 生成视频
      tags:
      - Video
  /v1/video/generations/{task_id}:
    get:
      consumes:
      - application/json
      description: 根据任务ID查询视频生成任务的状态和结果
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.VideoTaskResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询视频
      tags:
      - Video
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"